chore(product-components): move CoinLogo to product component
