+++ packages/product-components/src/components/CoinLogo/CoinLogo.tsx
