+++ packages/product-components/src/components/CoinLogo/coins.ts
