+++ packages/product-components/tsconfig.json
@@ -16,6 +16,9 @@
         {
             "path": "../../suite-common/validators"
         },
+        {
+            "path": "../../suite-common/wallet-config"
+        },
         { "path": "../components" },
         { "path": "../connect" },
         { "path": "../dom-utils" },
