+++ packages/suite/src/components/suite/modals/ReduxModal/UserContextModal/AdvancedCoinSettingsModal/AdvancedCoinSettingsModal.tsx
@@ -1,10 +1,11 @@
 import styled from 'styled-components';
-import { CoinLogo, variables } from '@trezor/components';
+import { variables } from '@trezor/components';
 import { Modal, Translation } from 'src/components/suite';
 import { networks, NetworkSymbol } from '@suite-common/wallet-config';
 import { CustomBackends } from './CustomBackends/CustomBackends';
 import { getCoinLabel } from 'src/utils/suite/getCoinLabel';
 import { useSelector } from 'src/hooks/suite';
+import { CoinLogo } from '@trezor/product-components';
 
 const Section = styled.div`
     display: flex;
