+++ packages/suite/src/components/suite/modals/ReduxModal/UserContextModal/DisableTorModal.tsx
@@ -1,7 +1,7 @@
 import { useState } from 'react';
 import styled from 'styled-components';
 
-import { Button, H3, Paragraph, CoinLogo, variables } from '@trezor/components';
+import { Button, H3, Paragraph, variables } from '@trezor/components';
 import { getTitleForNetwork } from '@suite-common/wallet-utils';
 import { UserContextPayload } from '@suite-common/suite-types';
 import { blockchainActions } from '@suite-common/wallet-core';
@@ -11,6 +11,7 @@ import { isOnionUrl } from 'src/utils/suite/tor';
 import { useCustomBackends } from 'src/hooks/settings/backends';
 import { NetworkSymbol } from '@suite-common/wallet-config';
 import { AdvancedCoinSettingsModal } from 'src/components/suite/modals';
+import { CoinLogo } from '@trezor/product-components';
 
 const BackendRowWrapper = styled.div`
     display: flex;
