+++ packages/suite/src/components/wallet/WalletLayout/AccountTopPanel/AccountTopPanel.tsx
@@ -3,13 +3,14 @@ import { forwardRef } from 'react';
 
 import { spacingsPx } from '@trezor/theme';
 import { NetworkSymbol } from '@suite-common/wallet-config';
-import { CoinLogo, SkeletonCircle, SkeletonRectangle } from '@trezor/components';
+import { SkeletonCircle, SkeletonRectangle } from '@trezor/components';
 
 import { FormattedCryptoAmount, AmountUnitSwitchWrapper } from 'src/components/suite';
 import { useSelector } from 'src/hooks/suite';
 import { FiatHeader } from 'src/components/wallet/FiatHeader';
 import { selectLocalCurrency } from 'src/reducers/wallet/settingsReducer';
 import { useFiatFromCryptoValue } from 'src/hooks/suite/useFiatFromCryptoValue';
+import { CoinLogo } from '@trezor/product-components';
 
 export const ACCOUNT_INFO_HEIGHT = 80;
